+++ web/src/settings_invites.ts
@@ -4,17 +4,21 @@ import {z} from "zod";
 import render_settings_resend_invite_modal from "../templates/confirm_dialog/confirm_resend_invite.hbs";
 import render_settings_revoke_invite_modal from "../templates/confirm_dialog/confirm_revoke_invite.hbs";
 import render_admin_invites_list from "../templates/settings/admin_invites_list.hbs";
+import render_edit_invite_user_modal from "../templates/settings/edit_invite_user_modal.hbs";
 
 import * as blueslip from "./blueslip";
 import * as channel from "./channel";
 import * as confirm_dialog from "./confirm_dialog";
+import * as dialog_widget from "./dialog_widget";
 import {$t, $t_html} from "./i18n";
+import {beforeSend, get_invite_streams} from "./invite";
 import * as ListWidget from "./list_widget";
 import * as loading from "./loading";
 import * as people from "./people";
 import * as settings_config from "./settings_config";
 import * as settings_data from "./settings_data";
 import {current_user, realm} from "./state_data";
+import * as stream_data from "./stream_data";
 import * as timerender from "./timerender";
 import * as ui_report from "./ui_report";
 import * as util from "./util";
@@ -35,6 +39,7 @@ export const invite_schema = z.intersection(
         z.object({
             is_multiuse: z.literal(true),
             link_url: z.string(),
+            stream_ids: z.array(z.number()),
         }),
     ]),
 );
@@ -48,12 +53,20 @@ type Invite = z.output<typeof invite_schema> & {
     img_src?: string;
 };
 
-const meta = {
+type Meta = {
+    loaded: boolean;
+    invites: Invite[];
+    invite_id?: number;
+};
+
+const meta: Meta = {
     loaded: false,
+    invites: [],
 };
 
 export function reset(): void {
     meta.loaded = false;
+    meta.invites = [];
 }
 
 function failed_listing_invites(xhr: JQuery.jqXHR): void {
@@ -88,6 +101,7 @@ function populate_invites(invites_data: {invites: Invite[]}): void {
 
     add_invited_as_text(invites_data.invites);
 
+    meta.invites = invites_data.invites;
     const $invites_table = $("#admin_invites_table").expectOne();
     ListWidget.create($invites_table, invites_data.invites, {
         name: "admin_invites_list",
@@ -105,6 +119,9 @@ function populate_invites(invites_data: {invites: Invite[]}): void {
             item.img_src = people.small_avatar_url_for_person(
                 people.get_by_user_id(item.invited_by_user_id),
             );
+            if (!settings_data.user_can_create_multiuse_invite()) {
+                item.disable_buttons = true;
+            }
             return render_admin_invites_list({invite: item});
         },
         filter: {
@@ -152,7 +169,7 @@ function do_revoke_invite({
 }): void {
     const modal_invite_id = $(".dialog_submit_button").attr("data-invite-id");
     const modal_is_multiuse = $(".dialog_submit_button").attr("data-is-multiuse");
-    const $revoke_button = $row.find("button.revoke");
+    const $revoke_button = $row.find("button.revoke-invite");
 
     if (modal_invite_id !== invite_id || modal_is_multiuse !== is_multiuse) {
         blueslip.error("Invite revoking canceled due to non-matching fields.");
@@ -183,7 +200,7 @@ function do_revoke_invite({
 
 function do_resend_invite({$row, invite_id}: {$row: JQuery; invite_id: string}): void {
     const modal_invite_id = $(".dialog_submit_button").attr("data-invite-id");
-    const $resend_button = $row.find("button.resend");
+    const $resend_button = $row.find("button.resend-invite");
 
     if (modal_invite_id !== invite_id) {
         blueslip.error("Invite resending canceled due to non-matching fields.");
@@ -229,6 +246,52 @@ export function set_up(initialize_event_handlers = true): void {
     });
 }
 
+type GetInvitationData = {
+    csrfmiddlewaretoken: string | undefined;
+    invite_as: number;
+    stream_ids: string;
+};
+
+function get_invitation_data(): GetInvitationData {
+    const invite_as = Number.parseInt(String($("#invite_as").val()), 10);
+    const stream_ids: number[] = [];
+
+    $("#invite-stream-checkboxes input:checked").each(function () {
+        const stream_id = Number.parseInt(String($(this).val()), 10);
+        stream_ids.push(stream_id);
+    });
+
+    const data = {
+        csrfmiddlewaretoken: $('input[name="csrfmiddlewaretoken"]').attr("value"),
+        invite_as,
+        stream_ids: JSON.stringify(stream_ids),
+    };
+    return data;
+}
+
+function do_edit_invite(): void {
+    const $invite_status = $("#dialog_error");
+    const data = get_invitation_data();
+
+    void channel.patch({
+        url: "/json/invites/multiuse/" + meta.invite_id,
+        data,
+        beforeSend,
+        success() {
+            dialog_widget.close();
+        },
+        error(xhr) {
+            const arr = JSON.parse(xhr.responseText);
+            ui_report.message(arr.msg, $invite_status, "alert-warning");
+        },
+        complete() {
+            $("#edit-invite-form .dialog_submit_button").text($t({defaultMessage: "Save changes"}));
+            $("#edit-invite-form .dialog_submit_button").prop("disabled", false);
+            $("#edit-invite-form .dialog_cancel_button").prop("disabled", false);
+        },
+    });
+}
+
 export function on_load_success(
     invites_data: {invites: Invite[]},
     initialize_event_handlers: boolean,
@@ -238,7 +301,7 @@ export function on_load_success(
     if (!initialize_event_handlers) {
         return;
     }
-    $(".admin_invites_table").on("click", ".revoke", (e) => {
+    $(".admin_invites_table").on("click", ".revoke-invite", (e) => {
         // This click event must not get propagated to parent container otherwise the modal
         // will not show up because of a call to `close_active` in `settings.js`.
         e.preventDefault();
@@ -269,7 +332,77 @@ export function on_load_success(
         $(".dialog_submit_button").attr("data-is-multiuse", is_multiuse);
     });
 
-    $(".admin_invites_table").on("click", ".resend", (e) => {
+    $(".admin_invites_table").on("click", ".edit-invite", (e) => {
+        // This click event must not get propagated to parent container otherwise the modal
+        // will not show up because of a call to `close_active_modal` in `settings.js`.
+        e.stopPropagation();
+        e.preventDefault();
+
+        meta.invite_id = Number.parseInt(String($(e.currentTarget).attr("data-invite-id")), 10);
+        const html_body = render_edit_invite_user_modal({
+            invite_as_options: settings_config.user_role_values,
+            is_admin: current_user.is_admin,
+            is_owner: current_user.is_owner,
+            streams: get_invite_streams(),
+            notifications_stream: stream_data.get_new_stream_announcements_stream(),
+        });
+
+        function invite_user_modal_post_render(): void {
+            $("#edit-invite-form .dialog_submit_button").prop("disabled", true);
+            const initial_invite: Invite | undefined = meta.invites.find(
+                (invite) => invite.id === meta.invite_id && invite.is_multiuse,
+            );
+            if (initial_invite?.is_multiuse) {
+                for (const stream_id of initial_invite.stream_ids) {
+                    $(`[value=${stream_id}]`).prop("checked", true);
+                }
+                $("#invite_as").val(initial_invite.invited_as);
+            }
+
+            function state_unchanged(): boolean {
+                if (!initial_invite || !initial_invite.is_multiuse) {
+                    return true;
+                }
+                const initial_streams = [...initial_invite.stream_ids].sort();
+                let selected_streams: number[] = [];
+                $("#streams_to_add input:checked").each(function () {
+                    const stream_id = Number.parseInt(String($(this).val()), 10);
+                    selected_streams.push(stream_id);
+                });
+                selected_streams = selected_streams.sort();
+                return (
+                    selected_streams.length === initial_streams.length &&
+                    selected_streams.every((val, index) => val === initial_streams[index]) &&
+                    Number.parseInt(String($("#invite_as").val()), 10) === initial_invite.invited_as
+                );
+            }
+
+            $("#edit-invite-form").on("change", "input:checkbox, select", () => {
+                $("#edit-invite-form .dialog_submit_button").prop("disabled", state_unchanged());
+            });
+            $("#invite_check_all_button").on("click", () => {
+                $("#streams_to_add :checkbox").prop("checked", true);
+                $("#edit-invite-form .dialog_submit_button").prop("disabled", state_unchanged());
+            });
+            $("#invite_uncheck_all_button").on("click", () => {
+                $("#streams_to_add :checkbox").prop("checked", false);
+                $("#edit-invite-form .dialog_submit_button").prop("disabled", state_unchanged());
+            });
+        }
+
+        dialog_widget.launch({
+            html_heading: $t_html({defaultMessage: "Edit invite"}),
+            html_body,
+            html_submit_button: $t_html({defaultMessage: "Save changes"}),
+            id: "edit-invite-form",
+            loading_spinner: true,
+            on_click: do_edit_invite,
+            post_render: invite_user_modal_post_render,
+            help_link: "/help/invite-new-users#edit-a-reusable-invitation-link",
+        });
+    });
+
+    $(".admin_invites_table").on("click", ".resend-invite", (e) => {
         // This click event must not get propagated to parent container otherwise the modal
         // will not show up because of a call to `close_active` in `settings.js`.
         e.preventDefault();