+++ zerver/openapi/python_examples.py
@@ -344,6 +344,48 @@ def create_reusable_invitation_link(client: Client) -> None:
     validate_against_openapi_schema(result, "/invites/multiuse", "post", "200")
 
 
+@openapi_test_function("/invites/multiuse/{invite_id}:patch")
+def edit_reusable_invitation_link(client: Client) -> None:
+    request = {
+        "invite_expires_in_minutes": 60 * 24 * 10,  # 10 days
+        "invite_as": 400,
+        "stream_ids": [1, 8],
+    }
+    result = client.call_endpoint(url="/invites/multiuse", method="POST", request=request)
+
+    # {code_example|start}
+    # Edit reusable invitation link
+    invite_id = 1
+    request = {
+        "invite_as": 600,
+        "stream_ids": [7, 9],
+    }
+    result = client.call_endpoint(
+        url=f"/invites/multiuse/{invite_id}", method="PATCH", request=request
+    )
+    # {code_example|end}
+
+    validate_against_openapi_schema(result, "/invites/multiuse/{invite_id}", "patch", "200")
+
+
+@openapi_test_function("/invites/multiuse/{invite_id}:delete")
+def revoke_reusable_invitation(client: Client) -> None:
+    request = {
+        "invite_expires_in_minutes": 60 * 24 * 10,  # 10 days
+        "invite_as": 400,
+        "stream_ids": [1, 8],
+    }
+    result = client.call_endpoint(url="/invites", method="POST", request=request)
+
+    # {code_example|start}
+    # Revoke email invitation
+    invite_id = 1
+    result = client.call_endpoint(url=f"/invites/multiuse/{invite_id}", method="DELETE")
+    # {code_example|end}
+
+    validate_against_openapi_schema(result, "/invites/multiuse/{invite_id}", "delete", "200")
+
+
 @openapi_test_function("/invites/{invite_id}:delete")
 def revoke_email_invitation(client: Client) -> None:
     request = {
@@ -1722,6 +1764,8 @@ def test_invitations(client: Client) -> None:
     send_invitations(client)
     revoke_email_invitation(client)
     create_reusable_invitation_link(client)
+    revoke_reusable_invitation(client)
+    edit_reusable_invitation_link(client)
     get_invitations(client)