+++ api_docs/changelog.md
@@ -20,6 +20,14 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 9.0
 
+**Feature level 257**
+
+* [`GET /invites`](/api/get-invites): The stream_ids field has been added to
+  be returned for all invites.
+* [`PATCH /invites/multiuse/{invite_id}`](/api/edit-invite-link): Added support for
+  editing multiuse invites to users as admin or non admin with regards to role and
+  associated streams.
+
 **Feature level 256**
 
 * [`GET /events`](/api/get-events): Stream update events with a new